chore: remove sqllog and introduce backend structs in drivers
